requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
defusedxml==0.7.1
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
fonttools==4.58.4
fpdf2==2.8.3
greenlet==3.2.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
load-dotenv==0.1.0
MarkupSafe==3.0.2
mysqlclient==2.2.7
packaging==25.0
pillow==11.2.1
python-dotenv==1.1.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
